Move test code from samples directory to test directory
