+++ test/stream_builder_custom.cpp
