+++ test/stream_builder_custom.h
