+++ test/stream_custom.h
